tests/wpt/web-platform-tests/mathml/relations/css-styling/padding-border-margin/padding-003.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>padding</title>
<link rel="help" href="https://mathml-refresh.github.io/mathml-core/#layout-algorithms">
<meta name="assert" content="Verify that padding is taken into account.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/mathml/support/mathml-fragments.js"></script>
<script src="helper.js"></script>
<script>
  var epsilon = 1;

  setup({ explicit_done: true });
  window.addEventListener("load", runTests);

  function runTests() {

    for (tag in MathMLFragments) {
        if (!FragmentHelper.isEmpty(tag))
            continue;

        test(function() {
            var style = "padding-left: 10px; padding-right: 20px;  padding-top: 15px; padding-bottom: 25px;";
            var s = compareSizeWithAndWithoutStyle(tag, style);
            assert_approx_equals(s.width_delta, 10 + 20, epsilon, "padding left/right");
            assert_approx_equals(s.height_delta, 15 + 25, epsilon, "padding top/bottom");
        }, `Padding properties on ${tag}`);
    }

    done();
  }
</script>
</head>
<body>
  <div id="log"></div>
</body>
</html>
